Extract dictation pipeline into service-independent DictationRuntime
Pure, bug-for-bug refactor (#143 Phase 1): the record -> transcribe ->
clean -> deliver pipeline moves verbatim out of WhisperAccessibilityService
into DictationRuntime, a plain class holding a Context and a new minimal
RuntimeListener host seam. The service becomes the accessibility host:
a11y lifecycle, overlay ring UI, streaming-partial field writes, all text
injection / preview / undo / clipboard affordances, and the existing
AccessibilityTextDestination delivery.

Moved into DictationRuntime (names, comments, threading intact):
- RecordingStateMachine ownership + onTap/startRecording/stopAndTranscribe/
  cancelTranscription state transitions
- RecordingEngine + per-recording SilenceAutoStopSession (#108) /
  AacEncoderSession (#109) with the #193 session-local release closures
- TranscriptionGuard / activeToken / watchdog / PipelineTimingSlot (#115,
  abandon-on-non-happy-path #192 semantics)
- onRecordingFinished / resolveLateRecordingOnMain (#66/#90) /
  startMaxDurationTranscription / continueTranscription (incl. #192
  min-duration floor), transcribeLocal (#132/#139), transcribeApi provider
  chain walk (M5 compressed-file lifetime), handleTranscriptionResult
  (junk gating #192, cleanup waterfall dispatch, cursor signature reset M3)
- transcriber slots + trim/warm lifecycle (#98), streaming recognizer slot
  + handleStreamingChunk decode, wakelock (M6), warm-up trio, orphan-PCM
  cleanup, model init/reload, correlationId minting

Host keeps and forwards: overlay/injection/preview/undo/history/suggestion
collection (all delivery-time state), companion statics (instance,
requestToggleRecording, currentRecordingState -> runtime.currentState()).
SystemControlsAccessibilityService inherits everything unchanged.

RuntimeListener callbacks are verbatim relocations of the code the old
inline call sites ran: ring state changes, pre-recording teardown, the
streaming handoff half of teardownStreamingPreview, and deliverText
(preview-before-inject branch + injectText exactly as before).

Refs #143
