+++ app/src/main/kotlin/com/trevornk/ramblr/WhisperAccessibilityService.kt
@@ -3,7 +3,6 @@ package com.trevornk.ramblr
 import android.accessibilityservice.AccessibilityService
 import android.app.KeyguardManager
 import android.content.ClipboardManager
-import android.content.ComponentCallbacks2
 import android.content.Context
 import android.content.Intent
 import android.content.res.ColorStateList
@@ -33,9 +32,6 @@ import android.widget.LinearLayout
 import android.widget.ProgressBar
 import android.widget.TextView
 import android.widget.Toast
-import androidx.work.WorkInfo
-import androidx.work.WorkManager
-import java.io.File
 import kotlin.concurrent.thread
 import kotlin.math.abs
 
@@ -74,7 +70,7 @@ const val DEFAULT_INJECT_FEEDBACK = "Copied to clipboard"
 
 /**
  * #115: wall-clock markers for one dictation's user-perceived latency, from the stop tap through
- * to the actual injection result -- see [WhisperAccessibilityService.pipelineTiming]'s kdoc for
+ * to the actual injection result -- see [DictationRuntime.pipelineTiming]'s kdoc for
  * the field's lifecycle. [stopTapAtMs] is the anchor every other field's elapsed-ms is measured
  * against when the [PipelineStage] benchmark line is finally built.
  */
@@ -166,8 +162,8 @@ open class WhisperAccessibilityService : AccessibilityService() {
          *  by e.g. [ModelDownloadWorker.notifyServiceModelReady] / LivePreviewActivity's
          *  `reloadStreamingModel()` call). Added for the github-flavor self-update install gate
          *  (Part 4, [SelfUpdateInstallGate]/[SelfUpdateInstallWorker]) to check "is dictation
-         *  idle right now" without exposing the private [stateMachine] field itself. */
-        fun currentRecordingState(): RecordingStateMachine.State? = instance?.stateMachine?.current()
+         *  idle right now" without exposing the private [runtime] field itself. */
+        fun currentRecordingState(): RecordingStateMachine.State? = instance?.runtime?.currentState()
 
         /**
          * Cross-component entry point for the Quick Settings tile (#127): [RamblrQsTileService]
@@ -196,7 +192,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
                 // un-hiding is treated as a precondition of starting rather than an option.
                 // See [shouldRestoreIconBeforeToggle] for the rule and why it's start-only.
                 if (shouldRestoreIconBeforeToggle(
-                        service.stateMachine.current(),
+                        service.runtime.currentState(),
                         IconHiddenState.isHidden(service),
                     )
                 ) {
@@ -246,7 +242,6 @@ open class WhisperAccessibilityService : AccessibilityService() {
         fun foregroundPackageNameOrNull(): String? = instance?.currentForegroundPackageName()
 
         private const val TAG = "PhoneWhisper"
-        private const val SAMPLE_RATE = 16000
         private const val BTN_DP = 44
         private const val PAD_DP = 10
         private const val MARGIN_DP = 8
@@ -265,8 +260,6 @@ open class WhisperAccessibilityService : AccessibilityService() {
 
         /** Hold the button this long while TRANSCRIBING to cancel (see overlay.setOnTouchListener). */
         private const val LONG_PRESS_CANCEL_MS = 500L
-        /** Backstop if no transcription/cleanup callback ever fires; covers transcription + cleanup callTimeouts. */
-        private const val WATCHDOG_TIMEOUT_MS = 400_000L
         /** Grace period before wiping the clipboard after a paste-style injection reads it. */
         private const val CLIPBOARD_CLEAR_DELAY_MS = 30_000L
         /** How long "undo last insertion" / "retry with raw text" stay available after an injection (#27). */
@@ -324,7 +317,100 @@ open class WhisperAccessibilityService : AccessibilityService() {
         }
     }
 
-    private val stateMachine = RecordingStateMachine()
+    /**
+     * The extracted dictation engine (#143 Phase 1): record -> transcribe -> clean -> deliver
+     * lives in [DictationRuntime]; this service is its accessibility host (overlay ring UI,
+     * a11y-based text injection, preview/undo/clipboard affordances). The listener below is the
+     * host half of the pre-extraction pipeline, verbatim: each callback body is exactly the code
+     * the corresponding call site ran inline before the extraction.
+     */
+    private val runtimeListener = object : RuntimeListener {
+        override fun onRecordingStartRequested() {
+            // A still-pending preview (#40) from the previous dictation shouldn't linger silently
+            // while a new one starts -- resolve it the same safe way a timeout would.
+            pendingPreview?.let { resolvePreview { p -> p.timeout() } }
+
+            endStreamingSession()
+            flushPendingStreamingHandoff()
+            lastBubblePartialText = null
+            lastBubblePartialAtMs = 0L
+        }
+
+        override fun onRecordingStartFailed() {
+            // End the streaming session so a failed recorder start doesn't leak the OnlineStream
+            // opened by beginSession() until the next recording (L11).
+            endStreamingSession()
+        }
+
+        override fun onRecordingStarted() {
+            setBusy(false)
+            // Deferred (Trevor-reported bug fix, see animateRingX's kdoc): with SingleTapRestoreToggle
+            // on, this exact call site can run while restoreFromPeek()'s animator is still mid-flight
+            // repositioning the same ring window from the SAME tap gesture. setAppearance()/
+            // startPulse() are a second, independent animation on the same button view; without this
+            // guard the two could interleave and the peek-restore slide would visibly stall or never
+            // finish, even though recording (and its own animation) started correctly.
+            deferUntilPeekAnimationDone {
+                setAppearance(COLOR_RECORDING)
+                startPulse()
+            }
+        }
+
+        override fun onEnterTranscribingUi() {
+            handler.post { stopPulse() }
+            setAppearance(COLOR_BUSY)
+            setBusy(true)
+        }
+
+        override fun onIdleUi() {
+            setBusy(false)
+            setAppearance(COLOR_IDLE)
+        }
+
+        override fun onStreamingTeardown() {
+            // The host half of the pre-extraction teardownStreamingPreview(): the node reference
+            // isn't recycled outright here but moved into [pendingStreamingHandoff] (#45) -- the
+            // final batch injection may still be pending behind a preview-before-inject commit
+            // (#40) at this point, and needs that reference to reconcile its tracked span once it
+            // actually runs. Called from every path back to IDLE and from onDestroy (which
+            // flushes it right after, since no injection will ever follow there); safe to call
+            // even when no session is active.
+            flushPendingStreamingHandoff()
+            pendingStreamingHandoff = streamingSession
+            streamingSession = null
+        }
+
+        override fun onStreamingPartial(text: String) {
+            maybeInjectPartial(text)
+        }
+
+        override fun deliverText(
+            text: String,
+            rawText: String?,
+            paidFallbackGroup: CleanupStepGroup?,
+            cleanupError: String?,
+            feedbackDurationMs: Long,
+        ) {
+            // rawText is non-null exactly when cleanup ran and succeeded -- the one branch that
+            // routed through preview-before-inject (#40) before the extraction.
+            if (rawText != null && PreviewBeforeInjectToggle.isEnabled(this@WhisperAccessibilityService)) {
+                beginPreview(rawText = rawText, candidateText = text, paidFallbackGroup = paidFallbackGroup)
+            } else {
+                injectText(
+                    text,
+                    rawText = rawText,
+                    feedbackDurationMs = feedbackDurationMs,
+                    paidFallbackGroup = paidFallbackGroup,
+                    cleanupError = cleanupError,
+                )
+            }
+        }
+
+        override fun foregroundPackageName(): String? = currentForegroundPackageName()
+    }
+
+    internal val runtime = DictationRuntime(this, runtimeListener)
+
     private var overlayView: FrameLayout? = null
     private var button: ImageView? = null
     private var spinner: ProgressBar? = null
@@ -343,91 +429,12 @@ open class WhisperAccessibilityService : AccessibilityService() {
      *  the actual fix for the SingleTapRestoreToggle race documented on [animateRingX]. Cleared
      *  (and invoked) from the animator's own end/cancel listener, never left to accumulate. */
     private var pendingPostPeekAction: Runnable? = null
-    /** The currently-armed transcription watchdog (#L14), held so it can be removed on normal
-     *  completion/teardown instead of one accumulating in the handler queue per dictation. */
-    private var watchdogRunnable: Runnable? = null
     private val idlePeekRunnable = Runnable { attemptAutoPeek() }
     // screenW/screenH as of the last time the ring's position was computed (#41) -- the baseline
     // handleScreenSizeChange() diffs against to detect a fold/unfold. Set alongside layoutParams
     // in showOverlay() and kept in sync every time it repositions the ring.
     private var lastScreenW = 0
     private var lastScreenH = 0
-    @Volatile private var recordingEngine: RecordingEngine? = null
-    /** Non-null only while a recording with silence-based auto-stop (#108, mode 1) active is in
-     *  progress -- see [startRecording]/[onRecordingFinished]. Zero-cost when the feature is off
-     *  or the VAD model isn't installed: this field simply stays null, no [SilenceAutoStopSession]
-     *  (and therefore no native [com.k2fsa.sherpa.onnx.Vad]) is ever created. */
-    @Volatile private var silenceAutoStopSession: SilenceAutoStopSession? = null
-    /** Non-null only while a recording with compressed-upload encoding (#109) active is in
-     *  progress -- see [startRecording]/[onRecordingFinished]. Zero-cost when [CompressedUploadToggle]
-     *  is off: this field simply stays null, no [AacEncoderSession] (and therefore no MediaCodec/
-     *  MediaMuxer) is ever created. */
-    @Volatile private var aacEncoderSession: AacEncoderSession? = null
-    private val guard = TranscriptionGuard()
-    /**
-     * Bounded partial wakelock across the recording+transcription window (M6 audit, 2026-08-26).
-     * Screen-off long dictations otherwise risk the CPU being throttled mid local decode (the
-     * multi-minute sherpa path), and OEM battery managers are quicker to silence the mic for an
-     * app holding no wakelock. Lazily created once (PowerManager.newWakeLock), then reused for
-     * every dictation; non-reference-counted so acquire/release are idempotent under the rapid
-     * cancel->restart races #193 documents -- a re-acquire simply refreshes the timeout, and a
-     * second release is a no-op instead of a RuntimeException.
-     *
-     * Deliberately minimal scope: a full foregroundServiceType="microphone" FGS migration is the
-     * robust fix for OEM mic silencing and is tracked as future work in #204 -- this is the
-     * bounded, low-risk slice of it.
-     */
-    private var transcriptionWakeLock: android.os.PowerManager.WakeLock? = null
-
-    /** Leak backstop for [transcriptionWakeLock]: recording caps at 10 min and a worst-case
-     *  local decode adds multi-minute tail, so 20 min comfortably covers one dictation while
-     *  guaranteeing the OS drops the lock even if every release path were somehow missed. Each
-     *  [acquireTranscriptionWakeLock] refreshes this window for the new dictation. */
-    private val TRANSCRIPTION_WAKELOCK_TIMEOUT_MS = 20 * 60 * 1000L
-
-    /** Acquires (or refreshes) the bounded partial wakelock for a dictation that just started
-     *  recording. Main thread only (called from [startRecording]). Any failure is swallowed:
-     *  a wakelock is an optimization for screen-off reliability, never worth failing a
-     *  dictation over. */
-    private fun acquireTranscriptionWakeLock() {
-        runCatching {
-            val lock = transcriptionWakeLock ?: (getSystemService(Context.POWER_SERVICE) as android.os.PowerManager)
-                .newWakeLock(android.os.PowerManager.PARTIAL_WAKE_LOCK, "ramblr:transcription")
-                .apply { setReferenceCounted(false) }
-                .also { transcriptionWakeLock = it }
-            lock.acquire(TRANSCRIPTION_WAKELOCK_TIMEOUT_MS)
-        }.onFailure { Log.w(TAG, "Couldn't acquire transcription wakelock", it) }
-    }
-
-    /** Releases the wakelock at a pipeline terminal state. isHeld-guarded (and the lock is
-     *  non-reference-counted), so calling this from multiple teardown paths -- [resetToIdle]'s
-     *  always-runs funnel and [onDestroy]'s belt-and-suspenders -- is safe. */
-    private fun releaseTranscriptionWakeLock() {
-        runCatching {
-            transcriptionWakeLock?.takeIf { it.isHeld }?.release()
-        }.onFailure { Log.w(TAG, "Couldn't release transcription wakelock", it) }
-    }
-    /** Per-process-launch unique prefix for [correlationIdFor] (real bug fix, 2026-07-17):
-     *  [guard]'s underlying token is an in-process [java.util.concurrent.atomic.AtomicInteger]
-     *  that always restarts at 1 on a fresh accessibility-service process (app update, OS memory
-     *  pressure eviction, device reboot, or the user toggling the service off/on) -- all of which
-     *  happen routinely on Android, not edge cases. Benchmark log analysis confirmed this
-     *  collides real, unrelated dictations recorded hours or days apart under the identical
-     *  "tok-1"/"tok-3"/etc correlationId, silently corrupting any per-dictation grouping done
-     *  against [BenchmarkLogger]'s data. This value is captured once when the service process is
-     *  actually created (not merely `onServiceConnected`, which can re-fire without a new
-     *  process) and folded into every correlationId this service mints, so IDs are unique across
-     *  restarts while [guard]/[TranscriptionGuard] itself is untouched -- its only real job is
-     *  same-process staleness detection via [TranscriptionGuard.isCurrent], which needs no
-     *  cross-process uniqueness at all. */
-    private val processLaunchId = System.currentTimeMillis()
-    private fun correlationIdFor(token: Int) = "tok-$processLaunchId-$token"
-    private val inFlightCall = InFlightCall()
-    private val cleanupCursor = CleanupWaterfallCursor()
-    /** Signature of the cleanup waterfall the [cleanupCursor]'s index was last recorded against, so
-     *  a chain edit in Settings (which reshapes the step list) resets the position-based cursor
-     *  rather than resuming at a now-different step (M3). */
-    private var lastCleanupWaterfallSignature: String? = null
 
     /** Registered in [registerNetworkCallback]; null when registration failed or after teardown. */
     private var networkCallback: ConnectivityManager.NetworkCallback? = null
@@ -436,18 +443,6 @@ open class WhisperAccessibilityService : AccessibilityService() {
      *  hides the instant the device locks and reappears right after unlock (see
      *  [overlayShouldBeVisible]'s keyguard doc). */
     private var screenStateReceiver: android.content.BroadcastReceiver? = null
-    @Volatile private var activeToken: Int = 0
-    /** #115: wall-clock markers for the current dictation's user-perceived, stop-tap-to-injection
-     *  timeline. Started the instant [activeToken] is minted (stop tap, or the max-duration
-     *  auto-stop path's synthetic "stop") and consumed once in [finishInjection] to write the
-     *  end-to-end [PipelineStage] benchmark line. Every non-happy-path exit -- no speech
-     *  detected, long-press cancel, watchdog timeout, transcription-chain failure -- abandons the
-     *  timeline instead of leaving it populated (H2, #192), so a later, unrelated injectText()
-     *  call (e.g. [onFeedbackTapped]'s raw-text retry) can never attribute stale timing to
-     *  itself. A single slot (not a per-token map) is safe here because exactly one dictation is
-     *  ever in flight at a time -- [guard]/[activeToken] already enforce that invariant
-     *  everywhere else in this class. */
-    private val pipelineTiming = PipelineTimingSlot()
     private val handler = Handler(Looper.getMainLooper())
     private val hideFeedback = Runnable {
         setFeedbackTouchable(false)
@@ -491,20 +486,6 @@ open class WhisperAccessibilityService : AccessibilityService() {
     // Scheduled post-injection clipboard restore (#5) — at most one in flight, cancelled if a new
     // injection starts or the service tears down first. See clipboardRestoreOutcomeFor.
     private var pendingClipboardRestore: Runnable? = null
-
-    // Local transcription engine (loaded lazily)
-    private val transcriberSlot = TranscriberSlot<LocalTranscriber> { it.release() }
-
-    // Streaming live-preview engine (#29) — only loaded when the opt-in setting is on and the
-    // streaming model is installed (see initStreamingModel/shouldUseStreamingPreview). Kept loaded
-    // across recordings like transcriberSlot; only its per-recording OnlineStream is torn down
-    // between dictations (see StreamingTranscriber.beginSession/endSession).
-    private val streamingTranscriberSlot = TranscriberSlot<StreamingTranscriber> { it.release() }
-
-    // Set by onTrimMemory (#98) when the transcriber slots were released under memory pressure;
-    // cleared once warmUpTranscribersIfTrimmed reloads them. Avoids reloading on every single
-    // recording start -- only after a real trim actually emptied the slots.
-    @Volatile private var transcribersTrimmed = false
 
     // Live-preview injection state for the current recording (#29), null when no partial has been
     // injected yet this recording or once the session has ended. Only touched on the main thread.
@@ -559,11 +540,11 @@ open class WhisperAccessibilityService : AccessibilityService() {
         registerAccessibilityButton()
         registerNetworkCallback()
         registerScreenStateReceiver()
-        thread { cleanupOrphanedRecordings() }
+        thread { runtime.cleanupOrphanedRecordings() }
         thread { ModelDownloader.pruneOrphanedModelDirs(this) }
         // Try to load local model in background
-        thread { initLocalModel() }
-        thread { initStreamingModel() }
+        thread { runtime.initLocalModel() }
+        thread { runtime.initStreamingModel() }
     }
 
     /**
@@ -613,7 +594,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
         val cm = getSystemService(ConnectivityManager::class.java) ?: return
         val callback = object : ConnectivityManager.NetworkCallback() {
             override fun onAvailable(network: Network) {
-                cleanupCursor.reset()
+                runtime.onDefaultNetworkChanged()
             }
         }
         try {
@@ -701,47 +682,18 @@ open class WhisperAccessibilityService : AccessibilityService() {
         instance = null
         unregisterNetworkCallback()
         unregisterScreenStateReceiver()
-        // If a recording is in progress, force the reader thread off RECORDING/TRANSCRIBING so
-        // it tears down the AudioRecord and discards buffered PCM instead of leaking the mic or
-        // silently continuing to record after the service appears off. reset() runs
-        // unconditionally (H2): even if recordingEngine reads null here, a late old-session
-        // handoff could have raced our read of it while a reader is still live, so we must always
-        // walk the shared state machine out of RECORDING/TRANSCRIBING.
-        stateMachine.reset()
-        recordingEngine?.awaitTeardown()
-        recordingEngine = null
-        // Belt-and-suspenders alongside the release already wired into startRecording's onFinished
-        // (#108): awaitTeardown above blocks until the reader thread's onFinished has run, so this
-        // is normally already null, but a stray VAD session must never survive service teardown.
-        silenceAutoStopSession?.release()
-        silenceAutoStopSession = null
-        // Belt-and-suspenders alongside the release already wired into startRecording's onFinished
-        // (#109): awaitTeardown above blocks until the reader thread's onFinished has run, so this
-        // is normally already null, but a stray encoder session must never survive service teardown
-        // and leave a partial .m4a temp file behind.
-        aacEncoderSession?.release()
-        aacEncoderSession = null
-        cancelWatchdog()
-        guard.cancel()
-        inFlightCall.cancel()
-        // M6 belt-and-suspenders: service teardown is terminal for any in-flight dictation, and
-        // resetToIdle() may never run for it. isHeld-guarded, so a no-dictation destroy is a no-op.
-        releaseTranscriptionWakeLock()
-        // The cancel above makes any in-flight local completion abort at its next piece check
-        // (#83), so the holder's daemon thread can close the cached ~1 GB cleanup model promptly
-        // without this main-thread teardown waiting on it (#74).
-        LocalCleanupModelHolder.releaseAsync()
-        teardownStreamingPreview()
+        // The runtime owns the recording/transcription teardown: state machine reset, reader
+        // teardown, stray session release, watchdog/guard/in-flight cancel, wakelock release,
+        // cleanup-model release, and streaming teardown -- in exactly the pre-extraction order
+        // (see [DictationRuntime.shutdown]).
+        runtime.shutdown()
         flushPendingStreamingHandoff()
         // Release the native transcriber recognizers too (M7): like onTrimMemory, replace(null) can
         // block on an in-flight transcription, so it runs off the main thread. Without this, a
         // service destroy/recreate in the same process (accessibility toggle off/on) leaves the old
         // instance's recognizers (batch model up to 465MB) resident alongside the new ones until
         // process death.
-        thread {
-            transcriberSlot.replace(null)
-            streamingTranscriberSlot.replace(null)
-        }
+        runtime.releaseTranscribersAsync()
         handler.removeCallbacks(expirePendingInjection)
         pendingInjection?.node?.recycle()
         pendingInjection = null
@@ -782,72 +734,15 @@ open class WhisperAccessibilityService : AccessibilityService() {
      */
     override fun onTrimMemory(level: Int) {
         super.onTrimMemory(level)
-        if (level >= ComponentCallbacks2.TRIM_MEMORY_RUNNING_LOW) {
-            LocalCleanupModelHolder.releaseAsync()
-            transcribersTrimmed = true
-            // replace(null) takes the slot's write lock, which blocks until any in-flight
-            // transcription (a full batch on a background thread, up to a 10-minute recording)
-            // finishes. onTrimMemory runs on the main thread, so doing this inline would ANR the
-            // accessibility service exactly when memory pressure coincides with active
-            // transcription. Hop to a background thread, mirroring reloadModel() (H3).
-            thread {
-                transcriberSlot.replace(null)
-                streamingTranscriberSlot.replace(null)
-            }
-        }
-    }
-
-    /** Deletes any temp PCM files left behind by a process death that skipped normal teardown. */
-    private fun cleanupOrphanedRecordings() {
-        cacheDir.listFiles { f -> f.name.startsWith("rec_") && f.name.endsWith(".pcm") }
-            ?.forEach { it.delete() }
-    }
-
-    private fun initLocalModel() {
-        val modelName = prefs().getString("model_name", "") ?: ""
-        val newTranscriber = if (modelName.isBlank()) {
-            // Auto-detect first available model
-            val models = LocalTranscriber.availableModels(this)
-            if (models.isNotEmpty()) {
-                Log.i(TAG, "Auto-detected model: ${models.first()}")
-                LocalTranscriber.create(this, models.first())
-            } else null
-        } else {
-            LocalTranscriber.create(this, modelName)
-        }
-        // Swap in the new transcriber, then release the old one — waiting for any transcription
-        // still in flight on it — so switching models never holds more than one native recognizer.
-        transcriberSlot.replace(newTranscriber)
-        if (newTranscriber != null) {
-            Log.i(TAG, "Local transcription ready")
-        } else {
-            Log.i(TAG, "No local model found, will use API")
-        }
+        runtime.onTrimMemory(level)
     }
 
     /** Reload local model (called from MainActivity when settings change) */
-    fun reloadModel() { thread { initLocalModel() } }
-
-    /**
-     * (Re)loads the streaming live-preview model (#29), gated on both the opt-in setting and the
-     * streaming model being installed (see [shouldUseStreamingPreview]). Called from MainActivity
-     * when either changes, mirroring [reloadModel]'s pattern for the offline model.
-     */
-    private fun initStreamingModel() {
-        val archive = prefs().getString("streaming_model_name", STREAMING_MODEL.archive) ?: STREAMING_MODEL.archive
-        val model = ModelDownloader.resolveActiveModel(STREAMING_MODEL_CATALOG, archive)
-        val enabled = shouldUseStreamingPreview(
-            settingEnabled = prefs().getBoolean("streaming_preview_enabled", false),
-            streamingModelInstalled = StreamingTranscriber.isAvailable(this, model)
-        )
-        val newTranscriber = if (enabled) StreamingTranscriber.create(this, model) else null
-        streamingTranscriberSlot.replace(newTranscriber)
-        Log.i(TAG, if (newTranscriber != null) "Streaming preview ready" else "Streaming preview unavailable")
-    }
+    fun reloadModel() { runtime.reloadModel() }
 
     /** Reload the streaming preview model (called from MainActivity when the toggle or the
      *  streaming model's install state changes). */
-    fun reloadStreamingModel() { thread { initStreamingModel() } }
+    fun reloadStreamingModel() { runtime.reloadStreamingModel() }
 
     // --- Overlay ---
 
@@ -944,7 +839,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
         // were never updated for this gesture (still holding whatever the previous gesture left).
         var consumedByPeekRestore = false
         val longPressAction = Runnable {
-            val action = overlayLongPressActionFor(stateMachine.current(), hasPendingInjection = pendingInjection != null)
+            val action = overlayLongPressActionFor(runtime.currentState(), hasPendingInjection = pendingInjection != null)
             val moved = abs(lastRawX - touchX) + abs(lastRawY - touchY)
             if (!shouldFireLongPress(action, movedPastThreshold = moved >= TAP_THRESHOLD_DP * dp)) return@Runnable
             longPressFired = true
@@ -967,7 +862,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
                     // edge-swipe gesture, or landing on the window's transparent padding) were
                     // indistinguishable in logcat -- both left "zero trace". This one line is the
                     // discriminator: if it's absent for a failed tap, the touch never reached us.
-                    Log.i(TAG, "Ring touch DOWN: peeked=$isPeeked state=${stateMachine.current()} x=${params.x}")
+                    Log.i(TAG, "Ring touch DOWN: peeked=$isPeeked state=${runtime.currentState()} x=${params.x}")
                     // A touch-down on a currently-peeked ring is a restore tap, not the start of a
                     // normal drag/tap/long-press gesture (Feature A): consume it here so it can
                     // never also fall through to onTap()'s mic-toggle on ACTION_UP.
@@ -993,7 +888,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
                     lastRawX = ev.rawX; lastRawY = ev.rawY
                     longPressFired = false
                     styleMenuOpenedByLongPress = false
-                    val pendingAction = overlayLongPressActionFor(stateMachine.current(), hasPendingInjection = pendingInjection != null)
+                    val pendingAction = overlayLongPressActionFor(runtime.currentState(), hasPendingInjection = pendingInjection != null)
                     if (pendingAction != OverlayLongPressAction.NONE) {
                         handler.postDelayed(longPressAction, LONG_PRESS_CANCEL_MS)
                     }
@@ -1329,7 +1224,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
     private fun attemptAutoPeek() {
         if (isPeeked) return
         if (!AutoPeekToggle.isEnabled(this)) { armIdlePeekTimer(); return }
-        if (!RingPeek.shouldAutoPeek(stateMachine.current())) { armIdlePeekTimer(); return }
+        if (!RingPeek.shouldAutoPeek(runtime.currentState())) { armIdlePeekTimer(); return }
         val overlay = overlayView ?: return
         // Mirrors applyOverlayVisibility's alpha-not-GONE fix (Pixel Fold display-transition
         // stall): hidden state is now expressed as alpha=0f, not View.GONE.
@@ -1522,7 +1417,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
      *  half-applied state. See [OverlayAppearance]'s own doc for the custom-icon-vs-color-controls
      *  and idle-only-fill-override product decisions this implements.
      *
-     *  Also sets [ImageView.contentDescription] from [stateMachine]'s current state (read
+     *  Also sets [ImageView.contentDescription] from the runtime's current state (read
      *  directly rather than derived from [stateColor], since every call site already maps 1:1
      *  from the same state) -- this is the app's single most-used interactive control (the
      *  floating overlay button that starts/stops/cancels dictation) and previously had no
@@ -1530,7 +1425,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
     private fun applyButtonAppearance(stateColor: Int) {
         val btn = button ?: return
         val appearance = OverlayAppearancePrefs.load(this)
-        btn.contentDescription = when (stateMachine.current()) {
+        btn.contentDescription = when (runtime.currentState()) {
             RecordingStateMachine.State.IDLE -> getString(R.string.overlay_button_idle_description)
             RecordingStateMachine.State.RECORDING -> getString(R.string.overlay_button_recording_description)
             RecordingStateMachine.State.TRANSCRIBING -> getString(R.string.overlay_button_transcribing_description)
@@ -1595,7 +1490,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
             btn.setPadding(pad, pad, pad, pad)
             btn.requestLayout()
 
-            val stateColor = when (stateMachine.current()) {
+            val stateColor = when (runtime.currentState()) {
                 RecordingStateMachine.State.IDLE -> COLOR_IDLE
                 RecordingStateMachine.State.RECORDING -> COLOR_RECORDING
                 RecordingStateMachine.State.TRANSCRIBING -> COLOR_BUSY
@@ -2130,7 +2025,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
         button?.let {
             it.animate().alpha(0.4f).setDuration(500).withEndAction {
                 it.animate().alpha(1f).setDuration(500).withEndAction {
-                    if (stateMachine.isRecording()) startPulse()
+                    if (runtime.isRecording()) startPulse()
                 }.start()
             }.start()
         }
@@ -2158,968 +2053,14 @@ open class WhisperAccessibilityService : AccessibilityService() {
     // --- State machine ---
 
     private fun onTap() {
-        when (stateMachine.current()) {
-            RecordingStateMachine.State.IDLE -> startRecording()
-            RecordingStateMachine.State.RECORDING -> stopAndTranscribe()
-            RecordingStateMachine.State.TRANSCRIBING -> {}
-        }
-    }
-
-    private fun startRecording() {
-        if (checkSelfPermission(android.Manifest.permission.RECORD_AUDIO)
-            != android.content.pm.PackageManager.PERMISSION_GRANTED) {
-            toast("Grant audio permission in Ramblr app"); return
-        }
-
-        warmUpLocalCleanupModelIfNeeded()
-        warmUpTranscribersIfTrimmed()
-        warmUpCloudConnectionsIfNeeded()
-
-        // A still-pending preview (#40) from the previous dictation shouldn't linger silently
-        // while a new one starts -- resolve it the same safe way a timeout would.
-        pendingPreview?.let { resolvePreview { p -> p.timeout() } }
-
-        endStreamingSession()
-        flushPendingStreamingHandoff()
-        lastBubblePartialText = null
-        lastBubblePartialAtMs = 0L
-        val streamingActive = streamingTranscriberSlot.get() != null
-        if (streamingActive) streamingTranscriberSlot.use { it.beginSession() }
-
-        // Silence-based auto-stop (#108, mode 1): additive and opt-in -- a session (and the
-        // native Vad it owns) is only ever created when the toggle is on AND the model is
-        // actually installed. Either condition being false means zero VAD instantiation and
-        // byte-for-byte identical recording behavior to before this feature existed.
-        val autoStopSession = if (SilenceAutoStopToggle.isEnabled(this)) {
-            ModelDownloader.vadModelFile(this, SILERO_VAD_MODEL)?.let { modelFile ->
-                try {
-                    SilenceAutoStopSession(
-                        modelFile = modelFile,
-                        thresholdMs = SilenceAutoStopThreshold.millisOrDefault(this),
-                        onSilenceThresholdExceeded = { handler.post { stopAndTranscribe() } },
-                    )
-                } catch (e: Exception) {
-                    Log.e(TAG, "Failed to start silence auto-stop VAD session", e)
-                    null
-                }
-            }
-        } else null
-        silenceAutoStopSession = autoStopSession
-
-        // Compressed-upload AAC encoding (#109): additive and opt-in -- a session (and the
-        // MediaCodec/MediaMuxer it owns) is only ever created when CompressedUploadToggle is on.
-        // Unlike #108's VAD, no model download or extra precondition is needed: AAC-LC is a
-        // built-in platform codec. Toggle off means zero AacEncoderSession instantiation and
-        // byte-for-byte identical recording behavior to before this feature existed.
-        val aacSession = if (CompressedUploadToggle.isEnabled(this)) {
-            try {
-                AacEncoderSession(cacheDir)
-            } catch (e: Exception) {
-                Log.e(TAG, "Failed to start AAC encoder session", e)
-                null
-            }
-        } else null
-        aacEncoderSession = aacSession
-
-        val onChunk: (ByteArray, Int) -> Unit = when {
-            streamingActive && autoStopSession != null && aacSession != null -> { buf, len ->
-                handleStreamingChunk(buf, len)
-                autoStopSession.onChunk(buf, len)
-                aacSession.onChunk(buf, len)
-            }
-            streamingActive && autoStopSession != null -> { buf, len ->
-                handleStreamingChunk(buf, len)
-                autoStopSession.onChunk(buf, len)
-            }
-            streamingActive && aacSession != null -> { buf, len ->
-                handleStreamingChunk(buf, len)
-                aacSession.onChunk(buf, len)
-            }
-            autoStopSession != null && aacSession != null -> { buf, len ->
-                autoStopSession.onChunk(buf, len)
-                aacSession.onChunk(buf, len)
-            }
-            streamingActive -> ::handleStreamingChunk
-            autoStopSession != null -> autoStopSession::onChunk
-            aacSession != null -> aacSession::onChunk
-            else -> { _, _ -> }
-        }
-
-        val engine = RecordingEngine(cacheDir, stateMachine)
-        val started = engine.start(
-            onFinished = { result ->
-                // H1 (#193): release the per-recording LOCALS captured by this closure, never the
-                // fields. A reader stalled in AudioRecord.read() can outlive a rapid
-                // cancel->restart; by the time its onFinished runs, the fields already hold the
-                // NEW recording's sessions, and releasing those would kill the new recording's
-                // silence auto-stop and finalize its AAC encoder mid-recording. The fields are
-                // only nulled when they still reference these exact instances -- the same
-                // identity-guard pattern onRecordingFinished uses for `recordingEngine === engine`.
-                autoStopSession?.release()
-                if (silenceAutoStopSession === autoStopSession) silenceAutoStopSession = null
-                val compressedFile = aacSession?.finish()
-                aacSession?.release()
-                if (aacEncoderSession === aacSession) aacEncoderSession = null
-                val finalResult = if (compressedFile != null) result.copy(compressedFile = compressedFile) else result
-                onRecordingFinished(engine, finalResult)
-            },
-            onChunk = onChunk
-        )
-        if (!started) {
-            // End the streaming session so a failed recorder start doesn't leak the OnlineStream
-            // opened by beginSession() above until the next recording (L11).
-            if (streamingActive) endStreamingSession()
-            autoStopSession?.release()
-            silenceAutoStopSession = null
-            aacSession?.release()
-            aacEncoderSession = null
-            toast("Couldn't start recording — mic busy?")
-            return
-        }
-
-        recordingEngine = engine
-        // M6: recording is genuinely live -- open the bounded wakelock window that spans
-        // recording + transcription; resetToIdle() (the terminal-state funnel) closes it.
-        // Placed after the `started` check so a failed recorder start never acquires.
-        acquireTranscriptionWakeLock()
-        setBusy(false)
-        // Deferred (Trevor-reported bug fix, see animateRingX's kdoc): with SingleTapRestoreToggle
-        // on, this exact call site can run while restoreFromPeek()'s animator is still mid-flight
-        // repositioning the same ring window from the SAME tap gesture. setAppearance()/
-        // startPulse() are a second, independent animation on the same button view; without this
-        // guard the two could interleave and the peek-restore slide would visibly stall or never
-        // finish, even though recording (and its own animation) started correctly.
-        deferUntilPeekAnimationDone {
-            setAppearance(COLOR_RECORDING)
-            startPulse()
-        }
-    }
-
-    /** Flips shared state; the reader thread notices, drains, tears down and hands off via [onRecordingFinished]. */
-    private fun stopAndTranscribe() {
-        if (!stateMachine.tryStartTranscribing()) return
-        activeToken = guard.start()
-        pipelineTiming.start(PipelineTiming(stopTapAtMs = System.currentTimeMillis(), correlationId = correlationIdFor(activeToken)))
-        armWatchdog(activeToken)
-        enterTranscribingUi()
-    }
-
-    private fun enterTranscribingUi() {
-        handler.post { stopPulse() }
-        setAppearance(COLOR_BUSY)
-        setBusy(true)
-    }
-
-    /**
-     * Pre-warms the local cleanup model (#95) the instant recording starts, so its cold GGUF
-     * load (mmap + first-touch page faults on a several-hundred-MB file) overlaps with the user
-     * still talking and the transcription that follows, instead of starting only once cleanup
-     * itself runs and eating into [CLEANUP_WATERFALL_HARD_CAP_MS]'s budget -- see
-     * [LocalCleanupModelHolder.warmUpAsync]'s kdoc for the failure mode this fixes.
-     *
-     * Deliberately checked (not unconditional): only bothers when cleanup is actually enabled
-     * and the configured waterfall would use LOCAL_LLM for at least one step, since otherwise
-     * this would resident-load a multi-hundred-MB model into memory for a dictation that will
-     * never touch it (e.g. cleanup off, or an all-cloud waterfall).
-     */
-    private fun warmUpLocalCleanupModelIfNeeded() {
-        if (!PostProcessingToggle.shouldRunCleanup(PostProcessingToggle.isEnabled(this))) return
-        val providerChain = ProviderChainStore.load(this)
-        if (!providerChain.usesLocalLlm()) return
-        val modelPath = ModelDownloader.localCleanupModelFile(this, LocalCleanupProvider.selectedModel(this))?.absolutePath
-            ?: return
-        LocalCleanupModelHolder.warmUpAsync(modelPath)
-    }
-
-    /**
-     * Reloads the batch (and, if enabled, streaming) transcriber if [onTrimMemory] released them
-     * under memory pressure (#98) -- mirrors [warmUpLocalCleanupModelIfNeeded]'s pre-warm timing:
-     * this runs the instant recording starts, so a reload (typically well under a second for
-     * these much-smaller-than-the-cleanup-LLM models) overlaps with the user still talking rather
-     * than adding perceived latency at transcription time. No-ops on the far more common case
-     * where no trim has happened since the last dictation.
-     */
-    private fun warmUpTranscribersIfTrimmed() {
-        if (!transcribersTrimmed) return
-        transcribersTrimmed = false
-        thread { initLocalModel() }
-        thread { initStreamingModel() }
-    }
-
-    /**
-     * Pre-warms DNS/TCP/TLS for whatever cloud host(s) this dictation would actually call (#100
-     * perceived-latency follow-up), the same "pay the cost while the user is still talking"
-     * timing [warmUpLocalCleanupModelIfNeeded]/[warmUpTranscribersIfTrimmed] already use. Reads
-     * the same [ProviderChainStore]/[CloudFeatureToggle] state the real transcription and cleanup
-     * call sites resolve against, so this only opens connections a real call could actually use.
-     *
-     * The [ProviderCredentialStore] gate (#168) is what makes that last sentence true: without
-     * it this warmed a host for a provider that has no key, which both real call paths refuse to
-     * contact, so a Local-mode user with cleanup off and no key still opened a TLS connection to
-     * api.openai.com on every dictation.
-     */
-    private fun warmUpCloudConnectionsIfNeeded() {
-        val chain = ProviderChainStore.load(this)
-        val transcriptionCandidates = ProviderChainRuntime.transcriptionCandidates(chain)
-        val cleanupChain = ProviderChainRuntime.effectiveChainForCleanup(chain, CloudFeatureToggle.cleanupEnabled(this))
-        val hosts = NetworkWarmup.hostsToWarm(transcriptionCandidates, cleanupChain) { kind ->
-            ProviderCredentialStore.isConfigured(this, kind)
-        }
-        NetworkWarmup.warmUpAsync(hosts)
+        runtime.onTap()
     }
 
     /** Long-press while TRANSCRIBING (see overlay.setOnTouchListener): abort the in-flight call and return to idle. */
     private fun cancelTranscription() {
-        if (stateMachine.current() != RecordingStateMachine.State.TRANSCRIBING) return
-        inFlightCall.cancel()
-        // H2 (#192): this dictation will never reach finishInjection, so drop its timeline now --
-        // otherwise the next unrelated injection would consume it as its own.
-        pipelineTiming.abandon()
-        resetToIdle()
-        toast("Transcription cancelled")
+        runtime.cancelTranscription()
     }
 
-    /** Backstop for a callback that never arrives (stalled socket, hung local model, etc). See #20. */
-    private fun armWatchdog(token: Int) {
-        // Remove any previously-armed watchdog so a fresh arm (e.g. the transcribe->cleanup handoff
-        // arms a second one) doesn't leave the first runnable sitting in the handler queue for its
-        // full 400s (L14). The runnable is held so resetToIdle/onDestroy can remove it on normal
-        // completion instead of letting one accumulate per dictation.
-        cancelWatchdog()
-        val runnable = Runnable {
-            watchdogRunnable = null
-            if (guard.isCurrent(token)) {
-                inFlightCall.cancel()
-                // H2 (#192): a timed-out dictation never reaches finishInjection; drop its
-                // timeline so a later unrelated injection can't consume it.
-                pipelineTiming.abandon()
-                resetToIdle()
-                toast("Transcription timed out")
-            }
-        }
-        watchdogRunnable = runnable
-        handler.postDelayed(runnable, WATCHDOG_TIMEOUT_MS)
-    }
-
-    private fun cancelWatchdog() {
-        watchdogRunnable?.let { handler.removeCallbacks(it) }
-        watchdogRunnable = null
-    }
-
-    /** Common teardown for every path back to IDLE: normal completion, cancel, or watchdog. */
-    private fun resetToIdle() {
-        cancelWatchdog()
-        guard.cancel()
-        activeToken = 0
-        // M6: the pipeline is terminal here -- injection finished, cancel, watchdog, or error
-        // (reset(msg) funnels here too) -- so the recording+transcription wakelock window ends.
-        // isHeld-guarded and non-reference-counted, so the paths where no dictation was running
-        // (e.g. preview-before-inject's early resetToIdle) are harmless no-ops or early releases
-        // that the next startRecording() re-acquires.
-        releaseTranscriptionWakeLock()
-        // #115: deliberately NOT abandoned here. resetToIdle() runs immediately after beginPreview()
-        // too (preview-before-inject, #40) -- well before the real injectText() call that resolves
-        // the preview, possibly seconds later on a timeout. Abandoning here would silently drop
-        // pipeline timing for every previewed dictation. Instead each non-happy-path exit
-        // (cancel, watchdog, no-speech, reset(msg)) calls pipelineTiming.abandon() itself (H2,
-        // #192), and the happy path consumes exactly once in finishInjection().
-        stateMachine.reset()
-        setBusy(false)
-        setAppearance(COLOR_IDLE)
-        teardownStreamingPreview()
-    }
-
-    /** Called on the reader thread once the AudioRecord has been drained and released. */
-    private fun onRecordingFinished(engine: RecordingEngine, result: RecordingEngine.Result) {
-        // Clear the reference only if it still points at THIS engine (H2): a late old-session
-        // handoff must not null out the *new* engine's reference, or onDestroy would skip its
-        // teardown and leave the new reader keeping the mic hot after the service appears off.
-        if (recordingEngine === engine) recordingEngine = null
-        if (result.discarded) return // service destroyed / recording forced off — nothing to transcribe
-
-        // #115: mark stop-tap -> reader-drain/PCM-handoff the instant it actually happens, on
-        // this reader thread, rather than waiting for the main-thread token resolution below --
-        // resolveLateRecordingOnMain's hop can add a real, variable delay that would otherwise be
-        // silently folded into this measurement.
-        pipelineTiming.markDrained(System.currentTimeMillis())
-
-        val token = when {
-            activeToken != 0 && guard.isCurrent(activeToken) -> activeToken
-            result.stopReason == RecordingEngine.StopReason.MAX_DURATION -> {
-                startMaxDurationTranscription(result)
-                return
-            }
-            else -> {
-                // No valid token from this thread's view — but that view can be stale (#66): the
-                // stop tap mints the token on the main thread *after* the RECORDING->TRANSCRIBING
-                // CAS this reader thread reacted to, so deciding to discard here could silently
-                // drop a real dictation. Resolve on the main thread instead, where token/state
-                // are authoritative — mirroring startMaxDurationTranscription's pattern.
-                handler.post { resolveLateRecordingOnMain(result) }
-                return
-            }
-        }
-        continueTranscription(result, token)
-    }
-
-    /** Main-thread resolution for a recording that finished without a token (#66/#90) — see
-     *  [resolveLateRecording] for the decision table. */
-    private fun resolveLateRecordingOnMain(result: RecordingEngine.Result) {
-        val token = activeToken
-        when (resolveLateRecording(token, guard.isCurrent(token), stateMachine.current())) {
-            LateRecordingResolution.CONTINUE_TRANSCRIPTION -> thread { continueTranscription(result, token) }
-            LateRecordingResolution.DISCARD -> {
-                result.pcmFile?.delete()
-                result.compressedFile?.delete()
-            }
-            LateRecordingResolution.DISCARD_AND_RESET -> {
-                // Mic error mid-recording (#90): the reader thread self-claimed TRANSCRIBING but
-                // no transcription will run and no watchdog was armed — without this reset the
-                // overlay stays stuck in TRANSCRIBING with taps as no-ops forever.
-                result.pcmFile?.delete()
-                result.compressedFile?.delete()
-                reset(result.errorMessage?.let { "Recording error: $it" } ?: "Recording stopped unexpectedly")
-            }
-        }
-    }
-
-    /**
-     * The reader thread can hit the duration cap without a UI tap, so no token exists yet. Mint
-     * that token on the main thread with a fresh state check; if the user/service already reset to
-     * IDLE, discard the temp file instead of resurrecting a cancelled transcription.
-     */
-    private fun startMaxDurationTranscription(result: RecordingEngine.Result) {
-        handler.post {
-            if (stateMachine.current() != RecordingStateMachine.State.TRANSCRIBING || activeToken != 0) {
-                result.pcmFile?.delete()
-                result.compressedFile?.delete()
-                return@post
-            }
-            val token = guard.start()
-            activeToken = token
-            // #115: max-duration auto-stop has no real user "stop tap" -- the cap itself is the
-            // trigger -- so anchor the timeline here instead, at the moment this app-side decided
-            // the dictation is over. drainAtMs is set to the same instant since the drain already
-            // happened (this whole branch runs after onRecordingFinished already returned) before
-            // any pipelineTiming existed to record it against.
-            val nowMs = System.currentTimeMillis()
-            pipelineTiming.start(PipelineTiming(stopTapAtMs = nowMs, correlationId = correlationIdFor(token), drainAtMs = nowMs))
-            armWatchdog(token)
-            enterTranscribingUi()
-            toast("Recording limit reached (10 min) — transcribing…")
-            thread { continueTranscription(result, token) }
-        }
-    }
-
-    private fun continueTranscription(result: RecordingEngine.Result, token: Int) {
-        val file = result.pcmFile
-        if (file == null) { reset("No audio captured"); return }
-        if (result.errorMessage != null) Log.e(TAG, "Recording ended with error: ${result.errorMessage}")
-
-        // M3a (#192): a recording under the minimum duration floor (~300ms of PCM) cannot contain
-        // usable speech -- discard it with the existing "No speech detected" UX *before* paying
-        // for a full local decode or cloud upload that would only ever produce a blank transcript.
-        if (isBelowMinimumDuration(file.length(), SAMPLE_RATE)) {
-            Log.i(TAG, "Recording below ${MIN_RECORDING_DURATION_MS}ms floor (${file.length()} bytes); discarding")
-            file.delete()
-            result.compressedFile?.delete()
-            reset("No speech detected")
-            return
-        }
-
-        val useLocal = prefs().getBoolean("use_local", true)
-        val allowCloudFallback = DictationModeToggle.allowCloudFallback(this)
-        // #109: the compressed .m4a (when the toggle was on and AacEncoderSession's encode
-        // succeeded) only ever matters to a cloud upload -- the local transcriber reads raw PCM
-        // samples directly, never a file upload, so it's discarded (never uploaded) on that path.
-        val compressedFile = result.compressedFile
-
-        when {
-            useLocal && transcriberSlot.get() != null -> {
-                compressedFile?.delete()
-                transcribeLocal(file, token)
-            }
-            // #98 UX follow-up: previously this fell straight through to transcribeApi() whenever
-            // the local model wasn't loaded yet, regardless of *why* -- including the real-world
-            // case of a fresh install where onboarding's model download is still in progress.
-            // Since a local-mode user typically never entered a cloud API key, that produced a
-            // confusing "Set API key in Ramblr app" error with no connection to the actual cause,
-            // making the whole dictation feature look silently broken during exactly the moment
-            // (onboarding's "Try it out" step) a new user is forming their first impression.
-            // Give an honest, specific message instead of misrouting to a path that was never
-            // configured -- unless #100's "fall back to cloud if on-device fails" toggle is on,
-            // in which case a not-yet-downloaded local model is exactly the case that toggle
-            // exists for.
-            useLocal && allowCloudFallback -> transcribeApi(file, token, compressedFile)
-            useLocal -> {
-                compressedFile?.delete()
-                file.delete()
-                reset("Local model still downloading — try again once it finishes")
-            }
-            else -> transcribeApi(file, token, compressedFile)
-        }
-    }
-
-    /**
-     * Current WorkManager state of [model]'s download, or null when no work has ever been
-     * enqueued for it (#139).
-     *
-     * Queried synchronously because [transcribeLocal] already runs on its own background thread
-     * -- never call this from the main thread. Any failure resolves to null, which makes
-     * [VadModelProvisioning.shouldFetch] fall through to enqueueing; `ExistingWorkPolicy.KEEP`
-     * then collapses the duplicate, so the safe direction is preserved.
-     */
-    private fun vadDownloadStateFor(model: Model): WorkInfo.State? = try {
-        WorkManager.getInstance(this)
-            .getWorkInfosForUniqueWork(ModelDownloadWorker.workName(model.archive))
-            .get()
-            .firstOrNull { !it.state.isFinished }
-            ?.state
-    } catch (e: Exception) {
-        Log.w(TAG, "Couldn't read VAD download state; treating as not in flight", e)
-        null
-    }
-
-    /**
-     * Decodes the recording one VAD-detected speech segment at a time when the Silero VAD model
-     * is available (#132), falling back to the original whole-file decode when it isn't.
-     *
-     * The segmented path is what keeps peak memory bounded by utterance length: the old path read
-     * the take into one FloatArray and handed all of it to a single `OfflineStream`, which drove
-     * ~2GB RSS on a 5:52 dictation and got the process OOM-killed before any text or error could
-     * reach the user. Contrary to this method's previous doc comment, that FloatArray was never
-     * the dominant allocation (~22MB for a 6-minute take) -- the single-shot native decode was.
-     *
-     * The fallback is deliberate rather than a hard failure: the VAD model is an optional ~1-2MB
-     * download today (see [SILERO_VAD_MODEL]), so a user who has never enabled silence auto-stop
-     * would otherwise lose local transcription entirely. Short takes are unaffected either way.
-     * Since #139 this path also provisions that model itself, so the fallback is a transient
-     * state for local-transcription users rather than a permanent one.
-     *
-     * PCM lifetime (M5 audit, 2026-08-26): [file] is deleted here on success (the transcript
-     * exists, nothing downstream needs audio) and on failure of the *direct* pure-local path
-     * ([onChainFailure] null -- nothing could retry, so keeping it would only leak cache). When
-     * the provider-chain walk calls this as a candidate it passes [onChainFailure], taking the
-     * failure-path PCM lifetime for itself: the file is kept alive so the next candidate can
-     * still transcribe it, and [transcribeApi]'s `advanceOrGiveUp` deletes it once the chain is
-     * exhausted -- the [TranscriptionChain.shouldDeletePcm] contract.
-     */
-    private fun transcribeLocal(file: File, token: Int, onChainFailure: ((String) -> Unit)? = null) {
-        thread {
-            // Benchmark-log timing starts before the read/decode too, so a failure there (rare,
-            // but possible on a corrupt/truncated PCM file) still gets an honest latency instead
-            // of one measured from a t0 that never ran (GH #100 benchmark logger).
-            val benchmarkStartMs = System.currentTimeMillis()
-            try {
-                val vadModelFile = ModelDownloader.vadModelFile(this, SILERO_VAD_MODEL)
-                // #139: provision the VAD model for local transcription itself, rather than
-                // leaving it to the unrelated silence-auto-stop toggle (#108). Without this the
-                // segmented decode below is unreachable for anyone who never enabled that
-                // feature, silently pinning them to the unsegmented path #132 was filed to fix.
-                // Asynchronous by design: this take still falls back below, and a later one gets
-                // segmented decode once the ~644KB download lands.
-                if (VadModelProvisioning.shouldFetch(
-                        modelInstalled = vadModelFile != null,
-                        downloadInFlight = ModelDownloadWorker.isInFlight(
-                            vadDownloadStateFor(SILERO_VAD_MODEL)
-                        ),
-                    )
-                ) {
-                    Log.i(TAG, "VAD model missing — enqueueing download for segmented decode (#139)")
-                    ModelDownloadWorker.enqueue(this, SILERO_VAD_MODEL)
-                }
-                val vad = vadModelFile?.let { SherpaVadHandle.create(it) }
-
-                val t0 = System.currentTimeMillis()
-                val durationSeconds = (file.length() / 2 / SAMPLE_RATE).toInt()
-                val text = try {
-                    transcriberSlot.use { transcriber ->
-                        if (vad != null) {
-                            transcriber.transcribeSegmented(file, vad, SAMPLE_RATE)
-                        } else {
-                            Log.i(TAG, "VAD model unavailable — decoding unsegmented (#132)")
-                            transcriber.transcribe(PcmFileBuffer.readAsFloatArray(file), SAMPLE_RATE)
-                        }
-                    } ?: throw IllegalStateException("Local model was unloaded during transcription")
-                } finally {
-                    vad?.close()
-                }
-                // Success: the transcript exists, so no candidate (chain or otherwise) needs the
-                // audio anymore -- mirrors the delete-on-success the cloud candidates do (M5).
-                file.delete()
-                val ms = System.currentTimeMillis() - t0
-                Log.i(TAG, "Local transcription: ${ms}ms, ${durationSeconds}s audio")
-                BenchmarkLogger.log(
-                    context = this,
-                    correlationId = correlationIdFor(token),
-                    transcription = BenchmarkStage(
-                        provider = ProviderKind.LOCAL.name,
-                        model = localTranscriptionModelId(),
-                        latencyMs = System.currentTimeMillis() - benchmarkStartMs,
-                        success = true,
-                    ),
-                    rawTextLength = text.length,
-                )
-                QualityLogger.log(
-                    context = this,
-                    correlationId = correlationIdFor(token),
-                    transcription = QualityStage(
-                        provider = ProviderKind.LOCAL.name,
-                        model = localTranscriptionModelId(),
-                    ),
-                    rawText = text,
-                )
-
-                handleTranscriptionResult(text, token)
-            } catch (e: Exception) {
-                Log.e(TAG, "Local transcription failed", e)
-                BenchmarkLogger.log(
-                    context = this,
-                    correlationId = correlationIdFor(token),
-                    transcription = BenchmarkStage(
-                        provider = ProviderKind.LOCAL.name,
-                        model = localTranscriptionModelId(),
-                        latencyMs = System.currentTimeMillis() - benchmarkStartMs,
-                        success = false,
-                        // #138: this catch already had the reason and sent it only to logcat,
-                        // which rotates within hours while this file is the durable artifact.
-                        error = sanitizeError(e.toString()),
-                    ),
-                )
-                QualityLogger.log(
-                    context = this,
-                    correlationId = correlationIdFor(token),
-                    transcription = QualityStage(
-                        provider = ProviderKind.LOCAL.name,
-                        model = localTranscriptionModelId(),
-                    ),
-                )
-                if (onChainFailure != null) {
-                    // Chain-candidate path (M5): the PCM is deliberately NOT deleted -- the walk
-                    // owns its lifetime now (see kdoc above), so the next candidate can still
-                    // upload it. advanceOrGiveUp hops to main, re-checks the guard, and either
-                    // advances or deletes the audio and surfaces this error at chain exhaustion.
-                    onChainFailure("Local error: ${e.message}")
-                } else {
-                    // Direct pure-local path (continueTranscription): no chain walk owns the PCM
-                    // and nothing can retry without one, so delete it here and report honestly.
-                    // Cloud fallback for this path is only wired at the "model not loaded yet"
-                    // branch in continueTranscription -- a post-load failure this late is rare
-                    // and re-recording is cheap.
-                    file.delete()
-                    handler.post {
-                        if (!guard.isCurrent(token)) return@post // cancelled or watchdog already reset the UI
-                        toast("Local error: ${e.message}")
-                        resetToIdle()
-                    }
-                }
-            }
-        }
-    }
-
-    /** The on-device transcription model currently selected via [TranscriptionActivity] (a
-     *  separate model-picker system from [ProviderChainEntry.transcriptionModel] -- see that
-     *  field's kdoc). Read straight from the same "model_name" preference key
-     *  [ModelDownloader]/[TranscriptionActivity] use, so benchmark-log entries reflect the real
-     *  active local model rather than guessing at a chain entry that doesn't apply to LOCAL. */
-    private fun localTranscriptionModelId(): String = prefs().getString("model_name", "") ?: ""
-
-    /** The user's personal vocabulary terms (#26/#114), read from the same "custom_vocabulary_terms"
-     *  prefs key [BehaviorActivity] writes -- shared by both the cleanup-stage prompt interpolation
-     *  below and the transcription-stage prompt biasing (OpenAI/Gemini) in [transcribeApi]. */
-    private fun vocabularyTerms(): List<String> =
-        VocabularyTerms.parse(prefs().getString("custom_vocabulary_terms", VocabularyTerms.DEFAULT_SERIALIZED))
-
-    private fun transcribeApi(file: File, token: Int, compressedFile: File? = null) {
-        val chain = ProviderChainStore.load(this)
-        val allowLocalFallback = DictationModeToggle.allowLocalFallback(this)
-        val candidates = ProviderChainRuntime.transcriptionCandidates(chain, allowLocalFallback)
-        if (candidates.isEmpty()) {
-            file.delete()
-            compressedFile?.delete()
-            reset("No transcription provider configured")
-            return
-        }
-
-        // M5: which compressed .m4a copy (if any) the walk still holds. A LOCAL candidate never
-        // uploads, so its branch deletes the compressed copy and nulls this -- and now that LOCAL
-        // can *fail through* to a later cloud candidate (see its branch below), that candidate
-        // must see null (upload raw PCM) rather than a File pointing at deleted bytes. Only ever
-        // written before a candidate runs and read after its completion posts back, so the
-        // walk's existing thread handoffs (thread{}/handler.post) order every access.
-        var remainingCompressedFile: File? = compressedFile
-
-        fun attempt(index: Int) {
-            if (index >= candidates.size) {
-                file.delete()
-                remainingCompressedFile?.delete()
-                reset("Set API key in Ramblr app")
-                return
-            }
-            val entry = candidates[index]
-
-            // A candidate failed to produce a transcript (unusable, HTTP error, timeout, or empty).
-            // Fall through to the next configured candidate -- Gemini, or the on-device LOCAL floor
-            // -- instead of losing the dictation. Only once every candidate is exhausted do we
-            // delete the audio and reset with the last real error. The PCM file is kept alive
-            // across the whole walk precisely so a later candidate can still use it (#H1). Runs on
-            // whatever thread the failing candidate's callback used; hop to the main thread and
-            // re-check the guard first so a cancel/watchdog reset stops the walk.
-            fun advanceOrGiveUp(error: String?) {
-                handler.post {
-                    if (!guard.isCurrent(token)) {
-                        // Cancelled or watchdog already reset the UI: nothing more to try, but the
-                        // PCM is still ours to clean up.
-                        file.delete()
-                        remainingCompressedFile?.delete()
-                        return@post
-                    }
-                    if (TranscriptionChain.hasNextCandidate(index, candidates.size)) {
-                        Log.w(TAG, "Transcription candidate #$index (${entry.kind}) failed: ${error ?: "unusable"}; trying next candidate")
-                        // M2 (2026-08-25 audit): the guard re-check above belongs on main, but
-                        // attempt() itself must not run there -- a fallback candidate's request
-                        // prep is real work (a GEMINI candidate readBytes()es the whole PCM, up
-                        // to ~10MB, and base64s it to ~13MB inside GeminiTranscriberClient before
-                        // OkHttp ever sees it). Mirrors the thread { continueTranscription(...) }
-                        // hop resolveLateRecordingOnMain/startMaxDurationTranscription already
-                        // use: attempt(0) already runs on that reader thread, so everything
-                        // attempt() touches is exercised off-main today -- the thread{}/
-                        // handler.post handoffs order every access to the walk's shared state
-                        // (remainingCompressedFile, guard, candidates).
-                        thread { attempt(index + 1) }
-                    } else {
-                        file.delete()
-                        remainingCompressedFile?.delete()
-                        reset("Error: ${error ?: "transcription failed"}")
-                    }
-                }
-            }
-
-            val localLoaded = transcriberSlot.get() != null
-            val hasCredential = when (entry.kind) {
-                ProviderKind.OPENAI -> ProviderCredentialStore.get(this, ProviderKind.OPENAI).isNotBlank()
-                ProviderKind.GEMINI -> ProviderCredentialStore.get(this, ProviderKind.GEMINI).isNotBlank()
-                else -> false
-            }
-            if (TranscriptionChain.precheck(entry.kind, hasCredential, localLoaded) == TranscriptionChain.Precheck.SKIP) {
-                when (entry.kind) {
-                    ProviderKind.LOCAL ->
-                        // A LOCAL entry whose model hasn't loaded advances to the next candidate;
-                        // only if it's the last one do we surface "still downloading" (#H1).
-                        advanceOrGiveUp("Local model still downloading — try again once it finishes")
-                    else -> {
-                        Log.w(TAG, "Skipping transcription provider ${entry.kind}: not usable (no credential / not implemented)")
-                        attempt(index + 1)
-                    }
-                }
-                return
-            }
-
-            when (entry.kind) {
-                ProviderKind.OPENAI -> {
-                    val apiKey = ProviderCredentialStore.get(this, ProviderKind.OPENAI)
-                    Log.i(TAG, "Cloud transcription via ProviderChain provider=${entry.kind} (OpenAI audio/transcriptions)")
-                    val transcribeStartMs = System.currentTimeMillis()
-                    // Honor the entry's base-URL override and TRANSCRIPTION model (#101/#102: a
-                    // separate field from entry.model, which is the CLEANUP model -- OpenAI's
-                    // chat-completions models like "gpt-5.4-mini" can never serve
-                    // /v1/audio/transcriptions, which needs "whisper-1"/"gpt-4o-transcribe").
-                    // .ifBlank { null } (#104 audit finding): the current UI never persists a
-                    // literal "" here, but without this guard a hand-edited SharedPreferences
-                    // value or a future regression could send an empty-string model id straight
-                    // to OpenAI instead of falling back to DEFAULT_MODEL -- matches the same
-                    // guard already applied on the Gemini call site just below.
-                    // #114 parts 1/2: bias transcription itself toward the user's vocabulary, not
-                    // just the cleanup stage -- same terms already read below at the cleanup call
-                    // site (see vocabularyTerms()).
-                    // M5: capture the compressed copy this attempt actually uploads as a local
-                    // (the #193 locals-capture pattern) -- remainingCompressedFile can be nulled
-                    // by a LOCAL candidate, and this callback must report/delete what IT sent.
-                    val uploadCompressedFile = remainingCompressedFile
-                    TranscriberClient.transcribe(
-                        file, apiKey, inFlightCall,
-                        baseUrl = entry.baseUrlOverride ?: PostProcessor.DEFAULT_BASE_URL,
-                        model = entry.transcriptionModel?.ifBlank { null } ?: TranscriberClient.DEFAULT_MODEL,
-                        vocabularyTerms = vocabularyTerms(),
-                        compressedFile = uploadCompressedFile,
-                    ) { result ->
-                        val roundTripMs = System.currentTimeMillis() - transcribeStartMs
-                        Log.i(TAG, "OpenAI transcription HTTP round-trip took ${roundTripMs}ms")
-                        val success = result.text != null && result.text.isNotBlank()
-                        BenchmarkLogger.log(
-                            context = this,
-                            correlationId = correlationIdFor(token),
-                            transcription = BenchmarkStage(
-                                provider = entry.kind.name,
-                                model = entry.transcriptionModel?.ifBlank { null } ?: TranscriberClient.DEFAULT_MODEL,
-                                latencyMs = roundTripMs,
-                                success = success,
-                                compressedUpload = uploadCompressedFile != null,
-                                // #138: a provider error envelope yields blank text, so this
-                                // records success=false; without the reason a bad key, a rate
-                                // limit and a timeout are indistinguishable after logcat rotates.
-                                error = sanitizeError(result.error),
-                            ),
-                            rawTextLength = result.text?.length,
-                        )
-                        QualityLogger.log(
-                            context = this,
-                            correlationId = correlationIdFor(token),
-                            transcription = QualityStage(
-                                provider = entry.kind.name,
-                                model = entry.transcriptionModel?.ifBlank { null } ?: TranscriberClient.DEFAULT_MODEL,
-                            ),
-                            rawText = result.text,
-                        )
-                        if (success) {
-                            file.delete()
-                            uploadCompressedFile?.delete()
-                            handleTranscriptionResult(result.text, token)
-                        } else {
-                            advanceOrGiveUp(result.error ?: "empty transcript")
-                        }
-                    }
-                }
-                ProviderKind.LOCAL -> {
-                    Log.i(TAG, "Transcription via ProviderChain provider=${entry.kind}")
-                    // The local transcriber reads raw PCM, never an upload (#109) -- but null the
-                    // walk-level handle too (M5), so a cloud candidate reached via LOCAL's new
-                    // fail-through below uploads the raw PCM instead of a deleted .m4a path.
-                    remainingCompressedFile?.delete()
-                    remainingCompressedFile = null
-                    // M5: as a chain candidate, LOCAL's failure must behave like any cloud
-                    // candidate's -- keep the PCM alive and walk on via advanceOrGiveUp (which
-                    // deletes it at chain exhaustion) instead of transcribeLocal's direct-path
-                    // delete-and-reset. See transcribeLocal's PCM-lifetime kdoc and
-                    // TranscriptionChain.shouldDeletePcm.
-                    transcribeLocal(file, token, onChainFailure = { error -> advanceOrGiveUp(error) })
-                }
-                ProviderKind.GEMINI -> {
-                    val apiKey = ProviderCredentialStore.get(this, ProviderKind.GEMINI)
-                    // Gemini's inline-audio path buffers the recording ~4x in memory, so gate it by
-                    // size: above the threshold, fall through to the next candidate rather than risk
-                    // an OOM stacked on the resident STT/cleanup models (M6). The compressed .m4a
-                    // (#109) is always much smaller than the PCM/WAV it was encoded from, but the
-                    // size check still gates on the original PCM: it's the cheap, reliable proxy for
-                    // recording length already used everywhere else on this path, and a length-gated
-                    // recording that's small enough to compress successfully is smaller still once
-                    // encoded, so this can only ever be more permissive, never wrongly reject.
-                    if (!GeminiTranscriberClient.canInlineAudio(file.length())) {
-                        Log.w(TAG, "Recording too large for Gemini inline audio (${file.length()} bytes); trying next candidate")
-                        advanceOrGiveUp("Recording too large for Gemini transcription")
-                    } else {
-                        Log.i(TAG, "Cloud transcription via ProviderChain provider=${entry.kind} (Gemini generateContent audio)")
-                        val geminiModel = entry.transcriptionModel?.ifBlank { null } ?: GeminiTranscriberClient.DEFAULT_MODEL
-                        val geminiStartMs = System.currentTimeMillis()
-                        // M5: same locals-capture as the OpenAI branch above.
-                        val uploadCompressedFile = remainingCompressedFile
-                        GeminiTranscriberClient.transcribe(
-                            file, apiKey, geminiModel, inFlightCall,
-                            vocabularyTerms = vocabularyTerms(),
-                            compressedFile = uploadCompressedFile,
-                        ) { result ->
-                            val success = result.text != null && result.text.isNotBlank()
-                            BenchmarkLogger.log(
-                                context = this,
-                                correlationId = correlationIdFor(token),
-                                transcription = BenchmarkStage(
-                                    provider = entry.kind.name,
-                                    model = geminiModel,
-                                    latencyMs = System.currentTimeMillis() - geminiStartMs,
-                                    success = success,
-                                    compressedUpload = uploadCompressedFile != null,
-                                    // #138: same as the OpenAI path -- Gemini reports failure as
-                                    // an error envelope with blank text, not an exception.
-                                    error = sanitizeError(result.error),
-                                ),
-                                rawTextLength = result.text?.length,
-                            )
-                            QualityLogger.log(
-                                context = this,
-                                correlationId = correlationIdFor(token),
-                                transcription = QualityStage(
-                                    provider = entry.kind.name,
-                                    model = geminiModel,
-                                ),
-                                rawText = result.text,
-                            )
-                            if (success) {
-                                file.delete()
-                                uploadCompressedFile?.delete()
-                                handleTranscriptionResult(result.text, token)
-                            } else {
-                                advanceOrGiveUp(result.error ?: "empty transcript")
-                            }
-                        }
-                    }
-                }
-                ProviderKind.ANTHROPIC, ProviderKind.OMNIROUTE -> attempt(index + 1) // filtered out by capability; defensive only
-            }
-        }
-
-        attempt(0)
-    }
-
-    private fun handleTranscriptionResult(text: String?, token: Int) {
-        if (text.isNullOrBlank()) {
-            // H2 (#192): a no-speech dictation never reaches finishInjection -- abandon its
-            // timing so a later unrelated injection (e.g. feedback-bubble raw-text retry) can't
-            // consume it and write a garbage benchmark line.
-            pipelineTiming.abandon()
-            handler.post {
-                if (!guard.isCurrent(token)) return@post
-                toast("No speech detected")
-                resetToIdle()
-            }
-            return
-        }
-
-        // M3b (#192): a non-blank but content-free transcript (ASR hallucinations like "." or
-        // "you"-length fragments) has nothing for a cleanup model to improve -- running the full
-        // waterfall (network calls, LOCAL_LLM load) on it is pure waste. Inject it raw directly.
-        if (isJunkTranscript(text)) {
-            Log.i(TAG, "Transcript is content-free (len=${text.length}); skipping cleanup waterfall")
-            handler.post {
-                if (!guard.isCurrent(token)) return@post
-                injectText(text)
-                resetToIdle()
-            }
-            return
-        }
-
-        val usePostProcessing = PostProcessingToggle.shouldRunCleanup(PostProcessingToggle.isEnabled(this))
-
-        if (usePostProcessing) {
-            // Phase 3 (#95): "Use cloud for Cleanup" toggle on the new unified Cloud screen --
-            // applied here as a pure filter ahead of cleanup resolution so this call site is the
-            // only place that needs to know about it; everything downstream (cleanupWaterfallFor,
-            // shouldUseCleanupExecutor, processProviderChain) keeps operating on a plain
-            // ProviderChain exactly as Phase 2 verified.
-            val providerChain = ProviderChainRuntime.effectiveChainForCleanup(
-                ProviderChainStore.load(this), CloudFeatureToggle.cleanupEnabled(this), DictationModeToggle.allowLocalFallback(this)
-            )
-            val cleanupWaterfall = ProviderChainRuntime.cleanupWaterfallFor(providerChain)
-
-            // Reset the position-based waterfall cursor whenever the chain reshapes since the last
-            // cleanup (add/remove/reorder/toggle in Settings), so a cached raw index can't resume at
-            // a step that now occupies a different position -- which would skip newly-added free/
-            // local steps and mis-attribute success to the wrong step (M3).
-            val waterfallSignature = cleanupWaterfallSignature(cleanupWaterfall)
-            if (waterfallSignature != lastCleanupWaterfallSignature) {
-                cleanupCursor.reset()
-                lastCleanupWaterfallSignature = waterfallSignature
-            }
-
-            // A zero-step provider chain means the user explicitly removed every executable cleanup
-            // step: cleanup is disabled, so inject raw instead of falling back to any legacy store.
-            if (cleanupWaterfall.steps.isEmpty()) {
-                handler.post {
-                    if (!guard.isCurrent(token)) return@post
-                    injectText(text)
-                    resetToIdle()
-                }
-                return
-            }
-
-            // A single-OpenAI chain has exactly one required credential and no other step to fall
-            // through to, so a missing key is worth failing fast on here, before the executor makes
-            // a network call it can only fail. Real multi-step chains resolve credentials inside
-            // CleanupWaterfallExecutor and can fall through past an unconfigured cloud step to
-            // another provider (including LOCAL). Note this is only a pre-flight check -- since
-            // #105 every chain, this one included, is executed by CleanupWaterfallExecutor.
-            if (!ProviderChainRuntime.shouldUseCleanupExecutor(providerChain) &&
-                ProviderCredentialStore.get(this, ProviderKind.OPENAI).isBlank()) {
-                handler.post {
-                    if (!guard.isCurrent(token)) return@post
-                    toast("Post-processing needs API key. Using raw text.")
-                    injectText(text)
-                    resetToIdle()
-                }
-                return
-            }
-
-            val savedPrompt = prefs().getString("post_processing_prompt", PostProcessor.DEFAULT_PROMPT) ?: PostProcessor.DEFAULT_PROMPT
-            val perAppPersonaKey = if (PerAppPersonaToggle.isEnabled(this)) {
-                PerAppPersonaStore.personaKeyFor(this, currentForegroundPackageName())
-            } else {
-                null
-            }
-            val rawPrompt = perAppPersonaKey
-                ?.let { CleanupPersonas.promptForExplicitSelection(PersonaRegistry.resolve(this, it)) }
-                ?: savedPrompt
-            val vocabulary = vocabularyTerms()
-            val prompt = PostProcessor.interpolateVocabulary(rawPrompt, vocabulary)
-
-            if (!guard.isCurrent(token)) return
-            Log.i(TAG, "Cleanup via ProviderChain entries=${providerChain.entries.map { it.kind }} executableSteps=${cleanupWaterfall.steps.map { it.group }}")
-            PostProcessor.processProviderChain(
-                text = text,
-                prompt = prompt,
-                chain = providerChain,
-                cursor = cleanupCursor,
-                cancelHolder = inFlightCall,
-                credentialLookup = { kind -> ProviderCredentialStore.get(this, kind) },
-                localModelPath = { ModelDownloader.localCleanupModelFile(this, LocalCleanupProvider.selectedModel(this))?.absolutePath },
-                localPrompt = LocalCleanupProvider.selectedSystemPrompt(this),
-                // #182 option 2: local cleanup applies the same terms as a deterministic
-                // post-pass over its output instead of in its prompt (which broke LFM2.5).
-                localVocabulary = vocabulary,
-                benchmarkContext = this,
-                benchmarkCorrelationId = correlationIdFor(token),
-            ) { result ->
-                handler.post {
-                    if (!guard.isCurrent(token)) return@post // cancelled or watchdog already reset the UI
-                    if (result.text != null && result.text.isNotBlank()) {
-                        val servingGroup = recordProviderChainCleanupSuccess(cleanupWaterfall)
-                        val paidFallbackGroup = servingGroup?.takeIf { it.isPaidFallback() }
-                        if (PreviewBeforeInjectToggle.isEnabled(this)) {
-                            beginPreview(rawText = text, candidateText = result.text, paidFallbackGroup = paidFallbackGroup)
-                        } else {
-                            injectText(result.text, rawText = text, paidFallbackGroup = paidFallbackGroup)
-                        }
-                    } else {
-                        // Log + surface the real failure reason (bad/missing key, HTTP status,
-                        // network error, etc.) instead of a generic "cleanup failed" that gives
-                        // the user and any future debugging nothing to go on (#98, Trevor hit
-                        // this directly: OpenAI key rejected/failed with zero visible reason).
-                        // result.error already carries this from PostProcessor.Result/
-                        // CleanupStepOutcome -- it was just being discarded here.
-                        val reason = result.error?.takeIf { it.isNotBlank() } ?: "unknown error"
-                        Log.w(TAG, "Cleanup failed, injecting raw text: $reason")
-                        // #175: pass the raw error, not a finished message -- injectText() builds
-                        // the user-facing notice once the injection method is known, so it can
-                        // state the truth about the clipboard and keep executor diagnostics
-                        // (nested prefixes, provider error bodies) out of a floating overlay.
-                        injectText(text, cleanupError = reason, feedbackDurationMs = 4000)
-                    }
-                    resetToIdle()
-                }
-            }
-        } else {
-            handler.post {
-                if (!guard.isCurrent(token)) return@post
-                injectText(text)
-                resetToIdle()
-            }
-        }
-    }
-
-    /** Marks whichever provider-chain cleanup step just served this cleanup call as healthy (#32), so the
-     *  Settings status dot reflects real usage, not just a Test-button press, and returns that
-     *  step's group so callers can attribute the result for dictation history's "paid fallback"
-     *  badge (#33). Which step that was isn't threaded through [PostProcessor.processProviderChain]'s
-     *  callback, so it's inferred from [cleanupCursor]'s last-known-good index instead -- safe to
-     *  read immediately after a success since [CleanupWaterfallCursor.recordSuccess] was just
-     *  called with "now", well inside the idle-expiry window [CleanupWaterfallCursor.startIndex]
-     *  checks. A total-waterfall failure is deliberately not attributed to any one step here,
-     *  since several steps may have failed for different reasons -- the Settings "Test" button is
-     *  the deterministic way to pin down which. */
-    private fun recordProviderChainCleanupSuccess(executableWaterfall: CleanupWaterfall): CleanupStepGroup? {
-        val succeededIndex = cleanupCursor.startIndex(System.currentTimeMillis())
-        val step = executableWaterfall.steps.getOrNull(succeededIndex) ?: return null
-        CleanupStepStatusStore.record(this, step, CleanupStepHealth.SUCCESS)
-        return step.group
-    }
 
     // --- Preview-before-inject (#40) ---
 
@@ -3175,33 +2116,10 @@ open class WhisperAccessibilityService : AccessibilityService() {
         }
     }
 
-    /** Safe to call from any thread: [resetToIdle] mutates main-thread-only state
-     *  (streamingSession/pendingStreamingHandoff, node recycling, the guard), and this is the
-     *  one entry point reachable from the RecordingEngine reader thread (fast tap-tap with zero
-     *  bytes captured, or missing API key) -- so it hops (#72). */
-    private fun reset(msg: String) {
-        toast(msg)
-        // H2 (#192): every reset(msg) call site is a terminal error/give-up exit (no audio, chain
-        // exhausted, recording error, ...) after which no finishInjection will run for this
-        // dictation -- abandon its timing so it can't be misattributed to a later injection.
-        pipelineTiming.abandon()
-        handler.post { resetToIdle() }
-    }
+    /** Safe to call from any thread: [DictationRuntime.resetToIdle] mutates main-thread-only
+     *  state -- so it hops (#72). Kept in the runtime now; the service has no reset of its own. */
 
     // --- Streaming preview (#29) ---
-
-    /**
-     * Runs on [RecordingEngine]'s reader thread (see [RecordingEngine.start]'s `onChunk`): converts
-     * the raw PCM chunk and feeds it to the streaming recognizer, then hops to the main thread only
-     * if there's a new hypothesis to potentially show. Decoding happens on every chunk regardless of
-     * whether the result ends up injected — [maybeInjectPartial] is what throttles actual UI/field
-     * updates, not this.
-     */
-    private fun handleStreamingChunk(buf: ByteArray, len: Int) {
-        val samples = PcmFileBuffer.bytesToFloatArray(buf, len)
-        val text = streamingTranscriberSlot.use { it.acceptChunk(samples, SAMPLE_RATE) } ?: return
-        handler.post { maybeInjectPartial(text) }
-    }
 
     /**
      * Injects (or replaces) the live partial preview in the focused field, throttled by
@@ -3217,7 +2135,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
      * for throttling purposes.
      */
     private fun maybeInjectPartial(text: String) {
-        if (!stateMachine.isRecording()) return // stale post after stop/cancel raced this
+        if (!runtime.isRecording()) return // stale post after stop/cancel raced this
         if (text.isBlank()) return
         val now = System.currentTimeMillis()
 
@@ -3318,20 +2236,6 @@ open class WhisperAccessibilityService : AccessibilityService() {
     private fun endStreamingSession() {
         streamingSession?.node?.recycle()
         streamingSession = null
-    }
-
-    /** Tears down everything about the current recording's live-preview session: the streaming
-     *  recognizer's per-recording stream, and the injected-node reference. The node reference isn't
-     *  recycled outright here but moved into [pendingStreamingHandoff] (#45) -- the final batch
-     *  injection may still be pending behind a preview-before-inject commit (#40) at this point, and
-     *  needs that reference to reconcile its tracked span once it actually runs. Called from every
-     *  path back to IDLE ([resetToIdle]) and from [onDestroy] (which flushes it right after, since
-     *  no injection will ever follow there); safe to call even when no session is active. */
-    private fun teardownStreamingPreview() {
-        flushPendingStreamingHandoff()
-        pendingStreamingHandoff = streamingSession
-        streamingSession = null
-        streamingTranscriberSlot.use { it.endSession() }
     }
 
     /** Discards [pendingStreamingHandoff] without attempting to reconcile it against any field --
@@ -3481,7 +2385,7 @@ open class WhisperAccessibilityService : AccessibilityService() {
         // nulled right after reading) so a later, unrelated retry-tap injectText() call (e.g.
         // onFeedbackTapped's raw-text retry, or a preview commit that runs well after the
         // original stop tap) never attributes this dictation's stale timing to itself.
-        pipelineTiming.consume()?.let { timing ->
+        runtime.pipelineTiming.consume()?.let { timing ->
             val nowMs = System.currentTimeMillis()
             BenchmarkLogger.log(
                 context = this,
